+++ catalog/app/containers/Queries/Athena/Athena.tsx
@@ -372,7 +372,7 @@ const useStyles = M.makeStyles((t) => ({
 
 function AthenaContainer() {
   const { queryExecutionId, workgroup } = Model.use()
-  const settings = CatalogSettings.use()
+  const beta = CatalogSettings.useBetaEnabled()
 
   const classes = useStyles()
   return (
@@ -398,7 +398,7 @@ function AthenaContainer() {
       {Model.hasData(workgroup.data) && (
         <div className={classes.content}>
           <div className={classes.section}>
-            {settings?.beta && <TabulatorTables />}
+            {beta && <TabulatorTables />}
             <QueryEditor.Form className={classes.form} />
           </div>
           {queryExecutionId ? (